refactor(auth): move jwt_secret() from views.rs to views/utils.rs
Enforce MF-1 convention: aggregator module files contain only pub mod
declarations and pub use re-exports. Update import paths in login.rs,
register.rs, and middleware.rs.
