+++ app/src/apps/auth/views/register.rs
@@ -7,9 +7,9 @@ use reinhardt::http::ViewResult;
 use reinhardt::post;
 use reinhardt::{BaseUser, Json, JwtAuth, Response, StatusCode};
 
+use super::utils::jwt_secret;
 use crate::apps::auth::models::User;
 use crate::apps::auth::serializers::{RegisterRequest, TokenResponse};
-use crate::apps::auth::views::jwt_secret;
 
 /// Register new user, persist to database, and return JWT token.
 #[post("/auth/register/", name = "auth_register", pre_validate = true)]
